requirements.txt
sagemath >= 1.2.5
# cython >= 0.25
